Combine tabular viz e2e specs (#43519)
