+++ e2e/test/scenarios/visualizations-tabular/column-shortcuts.cy.spec.ts
@@ -17,7 +17,7 @@ import {
   expectNoBadSnowplowEvents,
 } from "e2e/support/helpers";
 
-const { PEOPLE, PEOPLE_ID, ORDERS, ORDERS_ID } = SAMPLE_DATABASE;
+const { PEOPLE, PEOPLE_ID, ORDERS, ORDERS_ID, PRODUCTS } = SAMPLE_DATABASE;
 
 const DATE_CASES = [
   {
@@ -149,7 +149,11 @@ describeWithSnowplow("extract shortcut", () => {
       });
       openNotebook();
       getNotebookStep("expression").findByText("Year").click();
-      enterCustomColumnDetails({ formula: "year([Created At]) + 2" });
+      enterCustomColumnDetails({
+        name: "custom formula",
+        formula: "year([Created At]) + 2",
+        blur: true,
+      });
       popover().button("Update").click();
       visualize();
       cy.findByRole("gridcell", { name: "2,027" }).should("be.visible");
@@ -353,3 +357,145 @@ function extractColumnAndCheck({
     cy.findByRole("gridcell", { name: value }).should("be.visible");
   }
 }
+
+describeWithSnowplow("scenarios > visualizations > combine shortcut", () => {
+  function combineColumns({
+    columns,
+    example,
+    newColumn,
+    newValue,
+  }: {
+    columns: string[];
+    example: string;
+    newColumn: string;
+    newValue?: string;
+  }) {
+    const requestAlias = _.uniqueId("dataset");
+    cy.intercept("POST", "/api/dataset").as(requestAlias);
+    cy.findByLabelText("Add column").click();
+
+    popover().findByText("Combine columns").click();
+    for (const column of columns) {
+      selectColumn(column);
+    }
+
+    if (example) {
+      popover().findByTestId("combine-example").should("have.text", example);
+    }
+
+    popover().button("Done").click();
+
+    cy.wait(`@${requestAlias}`);
+
+    cy.findAllByRole("columnheader")
+      .last()
+      .should("have.text", newColumn)
+      .should("be.visible");
+
+    if (newValue) {
+      cy.findByRole("gridcell", { name: newValue }).should("be.visible");
+    }
+  }
+
+  function selectColumn(name: string) {
+    popover().findAllByText("Select a column...").first().click();
+    popover().last().findByText(name).click();
+  }
+
+  beforeEach(() => {
+    restore();
+    cy.signInAsNormalUser();
+    resetSnowplow();
+  });
+
+  afterEach(() => {
+    expectNoBadSnowplowEvents();
+  });
+
+  it("should be possible add a new column through the combine columns shortcut", () => {
+    createQuestion(
+      {
+        query: {
+          "source-table": PEOPLE_ID,
+          limit: 1,
+          fields: [
+            ["field", PEOPLE.ID, null],
+            ["field", PEOPLE.EMAIL, null],
+          ],
+        },
+      },
+      {
+        visitQuestion: true,
+      },
+    );
+
+    combineColumns({
+      columns: ["Email", "ID"],
+      newColumn: "Combined Email, ID",
+      example: "email@example.com12345",
+      newValue: "borer-hudson@yahoo.com1",
+    });
+
+    expectGoodSnowplowEvent({
+      event: "column_combine_via_plus_modal",
+      custom_expressions_used: ["concat"],
+      database_id: SAMPLE_DB_ID,
+    });
+  });
+
+  it("should allow combining columns when aggregating", function () {
+    createQuestion(
+      {
+        query: {
+          "source-table": ORDERS_ID,
+          limit: 1,
+          aggregation: [["count"]],
+          breakout: [
+            ["field", ORDERS.CREATED_AT, { "temporal-unit": "hour-of-day" }],
+          ],
+        },
+      },
+      {
+        visitQuestion: true,
+      },
+    );
+
+    cy.findByTestId("TableInteractive-root").should("exist");
+    combineColumns({
+      columns: ["Created At: Hour of day", "Count"],
+      newColumn: "Combined Created At: Hour of day, Count",
+      example: "2042-01-01 12:34:56.789 123",
+      newValue: "0 766",
+    });
+  });
+
+  it("should allow combining columns on a table with just breakouts", () => {
+    createQuestion(
+      {
+        query: {
+          "source-table": ORDERS_ID,
+          limit: 1,
+          breakout: [
+            ["field", ORDERS.CREATED_AT, { "temporal-unit": "hour-of-day" }],
+            [
+              "field",
+              PRODUCTS.CATEGORY,
+              { "base-type": "type/text", "source-field": ORDERS.PRODUCT_ID },
+            ],
+          ],
+        },
+      },
+      {
+        visitQuestion: true,
+      },
+    );
+
+    cy.findByTestId("TableInteractive-root").should("exist");
+    combineColumns({
+      columns: ["Created At: Hour of day", "Category"],
+      newColumn: "Combined Created At: Hour of day, Category",
+      example: "2042-01-01 12:34:56.789 text",
+      newValue: "0 Doohickey",
+    });
+  });
+});
